textrender: gameTextSetColor 83.85->100 (u8 params, not int)
The retail if-branch stores each arg directly with a bare `stb` into the
u8 color globals (no clrlwi mask), which only happens when the param is
already byte-width. Typing r/g/b/a as u8 emits the bare stb in the
draw-func branch and the clrlwi+stw zero-extend into the int command
slots in the else branch, matching retail exactly. All ~90 callers
(incl. gameui/gameloop/intersect/tumbleweedbush passing computed ints)
verified byte-identical.

+++ include/main/gametext_color_api.h
@@ -3,6 +3,6 @@
 
 #include "types.h"
 
-void gameTextSetColor(int r, int g, int b, int a);
+void gameTextSetColor(u8 r, u8 g, u8 b, u8 a);
 
 #endif /* MAIN_GAMETEXT_COLOR_API_H_ */
